fix:精简mermaid相关hooks
