+++ packages/x-markdown/src/hooks/index.ts
@@ -4,4 +4,3 @@ export * from './useMarkdown'
 export * from './useMermaid'
 export * from './usePlugins'
 export * from './useTheme'
-export * from './useMermaidZoom'
